+++ src/supervisor/network_commands.c
@@ -11,6 +11,7 @@
 
 #include "mac_mapper.h"
 #include "supervisor.h"
+#include "supervisor_utils.h"
 #include "sqlite_macconn_writer.h"
 #include "network_commands.h"
 
@@ -31,53 +32,6 @@
 
 #define ANALYSER_FILTER_FORMAT                                                 \
   "\"ether dst " MACSTR " or ether src " MACSTR "\""
-
-#ifdef WITH_CRYPTO_SERVICE
-int save_to_crypt(struct crypt_context *crypt_ctx, struct mac_conn_info *info) {
-  struct crypt_pair pair;
-
-  pair.key = info->id;
-  pair.value = info->pass;
-  pair.value_size = info->pass_len;
-
-  if (put_crypt_pair(crypt_ctx, &pair) < 0) {
-    log_trace("put_crypt_pair fail");
-    return -1;
-  }
-
-  return 0;
-}
-#endif
-
-bool save_mac_mapper(struct supervisor_context *context, struct mac_conn conn) {
-  if (!strlen(conn.info.id)) {
-    generate_radom_uuid(conn.info.id);
-  }
-
-  if (!put_mac_mapper(&context->mac_mapper, conn)) {
-    log_error("put_mac_mapper fail");
-    return false;
-  }
-
-#ifdef WITH_CRYPTO_SERVICE
-  if (save_to_crypt(context->crypt_ctx, &(conn.info)) < 0) {
-    log_error("save_to_crypt failure");
-    return false;
-  }
-
-  // Reset the plain password array so that it is not stored
-  // in plain form in the sqlite db
-  conn.info.pass_len = 0;
-  os_memset(conn.info.pass, 0, AP_SECRET_LEN);
-#endif
-
-  if (save_sqlite_macconn_entry(context->macconn_db, &conn) < 0) {
-    log_error("upsert_sqlite_macconn_entry fail");
-    return false;
-  }
-
-  return true;
-}
 
 void free_ticket(struct supervisor_context *context) {
   struct auth_ticket *ticket = context->ticket;
